Upd cam_cal and lens

+++ drivers/misc/mediatek/lens/common/inc/BU6424AF.h
@@ -11,11 +11,11 @@
 /* Structures */
 typedef struct {
 /* current position */
-	u32 u4CurrentPosition;
+	unsigned long u4CurrentPosition;
 /* macro position */
-	u32 u4MacroPosition;
+	unsigned long u4MacroPosition;
 /* Infiniti position */
-	u32 u4InfPosition;
+	unsigned long u4InfPosition;
 /* Motor Status */
 	bool bIsMotorMoving;
 /* Motor Open? */
@@ -33,11 +33,11 @@ typedef struct {
 /* H means "switch T and Q atomically" */
 #define BU6424AFIOC_G_MOTORINFO _IOR(BU6424AF_MAGIC, 0, stBU6424AF_MotorInfo)
 
-#define BU6424AFIOC_T_MOVETO _IOW(BU6424AF_MAGIC, 1, u32)
+#define BU6424AFIOC_T_MOVETO _IOW(BU6424AF_MAGIC, 1, unsigned long)
 
-#define BU6424AFIOC_T_SETINFPOS _IOW(BU6424AF_MAGIC, 2, u32)
+#define BU6424AFIOC_T_SETINFPOS _IOW(BU6424AF_MAGIC, 2, unsigned long)
 
-#define BU6424AFIOC_T_SETMACROPOS _IOW(BU6424AF_MAGIC, 3, u32)
+#define BU6424AFIOC_T_SETMACROPOS _IOW(BU6424AF_MAGIC, 3, unsigned long)
 
 #else
 #endif
